requirements.txt
Flask==2.3.2
#Flask==3.1.0
Flask-SocketIO==5.3.4
pymongo==4.4.1
flask_marshmallow==1.3.0
celery==5.4.0
pytest==7.4.3
pytest-mock==3.14.0
flask-jwt-extended==4.7.0
bcrypt==4.0.1
redis==5.0.0
flask-cors==4.0.0